+++ lib/src/commands/chat_command.dart
@@ -57,6 +57,14 @@ enum CommandType {
 
   /// Indicates that a [ChatCommand] can be executed by both Slash Commands and text messages.
   all,
+
+  /// Indicates that a [ChatCommand] should use the default type provided by [IOptions.options].
+  ///
+  /// If the default type provided by the options is itself [def], the behaviour is identical to
+  /// [all].
+  // TODO: Instead of having [def], make [ChatCommand.type] be a classical option
+  // ([ChatCommand.options.type]) and have it be inherited.
+  def,
 }
 
 mixin ChatGroupMixin implements IChatCommandComponent {
@@ -126,7 +134,7 @@ mixin ChatGroupMixin implements IChatCommandComponent {
     if (_childrenMap.containsKey(name)) {
       IChatCommandComponent child = _childrenMap[name]!;
 
-      if (child is ChatCommand && child.type != CommandType.slashOnly) {
+      if (child is ChatCommand && child.resolvedType != CommandType.slashOnly) {
         ChatCommand? found = child.getCommand(view);
 
         if (found == null) {
@@ -152,7 +160,8 @@ mixin ChatGroupMixin implements IChatCommandComponent {
 
   @override
   bool get hasSlashCommand => children.any((child) =>
-      (child is ChatCommand && child.type != CommandType.textOnly) || child.hasSlashCommand);
+      (child is ChatCommand && child.resolvedType != CommandType.textOnly) ||
+      child.hasSlashCommand);
 
   @override
   Iterable<CommandOptionBuilder> getOptions(CommandsPlugin commands) {
@@ -166,7 +175,7 @@ mixin ChatGroupMixin implements IChatCommandComponent {
           child.description,
           options: List.of(child.getOptions(commands)),
         ));
-      } else if (child is ChatCommand && child.type != CommandType.textOnly) {
+      } else if (child is ChatCommand && child.resolvedType != CommandType.textOnly) {
         options.add(CommandOptionBuilder(
           CommandOptionType.subCommand,
           child.name,
@@ -278,9 +287,25 @@ class ChatCommand
   /// commands executable only through Slash Commands, or only through text messages.
   ///
   /// You might also be interested in:
+  /// - [resolvedType], for getting the resolved type of this command.
   /// - [ChatCommand.slashOnly], for creating [ChatCommand]s with type [CommandType.slashOnly];
   /// - [ChatCommand.textOnly], for creating [ChatCommand]s with type [CommandType.textOnly].
   final CommandType type;
+
+  /// The resolved type of this [ChatCommand].
+  ///
+  /// If [type] is [CommandType.def], this will query the parent of this command for the default
+  /// type. Otherwise, [type] is returned.
+  ///
+  /// If [type] is [CommandType.def] and no parent provides a default type, [CommandType.def] is
+  /// returned.
+  CommandType get resolvedType {
+    if (type != CommandType.def) {
+      return type;
+    }
+
+    return resolvedOptions.defaultCommandType ?? CommandType.def;
+  }
 
   /// The function called to execute this command.
   ///
@@ -314,6 +339,9 @@ class ChatCommand
   /// - [checks] and [check], the equivalent for inherited checks.
   final List<AbstractCheck> singleChecks = [];
 
+  /// The types of the required and positional arguments of [execute], in the order they appear.
+  final List<Type> argumentTypes = [];
+
   @override
   final CommandOptions options;
 
@@ -337,7 +365,7 @@ class ChatCommand
     String description,
     Function execute, {
     List<String> aliases = const [],
-    CommandType type = CommandType.all,
+    CommandType type = CommandType.def,
     Iterable<IChatCommandComponent> children = const [],
     Iterable<AbstractCheck> checks = const [],
     Iterable<AbstractCheck> singleChecks = const [],
@@ -407,7 +435,7 @@ class ChatCommand
     this.execute,
     Type contextType, {
     this.aliases = const [],
-    this.type = CommandType.all,
+    this.type = CommandType.def,
     Iterable<IChatCommandComponent> children = const [],
     Iterable<AbstractCheck> checks = const [],
     Iterable<AbstractCheck> singleChecks = const [],
@@ -481,9 +509,15 @@ class ChatCommand
         throw CommandRegistrationError(
             'Command callback parameters must not have more than one UseConverter annotation');
       }
+      if (parametrer.metadata.where((element) => element.reflectee is Autocomplete).length > 1) {
+        throw CommandRegistrationError(
+            'Command callback parameters must not have more than one UseConverter annotation');
+      }
     }
 
     for (final argument in _arguments) {
+      argumentTypes.add(argument.type.reflectedType);
+
       Iterable<Name> names = argument.metadata
           .where((element) => element.reflectee is Name)
           .map((nameMirror) => nameMirror.reflectee)
@@ -632,7 +666,7 @@ class ChatCommand
 
   @override
   Iterable<CommandOptionBuilder> getOptions(CommandsPlugin commands) {
-    if (type != CommandType.textOnly) {
+    if (resolvedType != CommandType.textOnly) {
       List<CommandOptionBuilder> options = [];
 
       for (final mirror in _arguments) {
@@ -657,13 +691,17 @@ class ChatCommand
 
         choices ??= argumentConverter?.choices;
 
-        options.add(CommandOptionBuilder(
+        CommandOptionBuilder builder = CommandOptionBuilder(
           argumentConverter?.type ?? CommandOptionType.string,
           name,
           _mappedDescriptions[name]!.value,
           required: !mirror.isOptional,
           choices: choices?.toList(),
-        ));
+        );
+
+        argumentConverter?.processOptionCallback?.call(builder);
+
+        options.add(builder);
       }
 
       return options;
@@ -680,9 +718,9 @@ class ChatCommand
           'All child commands of chat groups or commands must implement IChatCommandComponent');
     }
 
-    if (type != CommandType.textOnly) {
+    if (resolvedType != CommandType.textOnly) {
       if (command.hasSlashCommand ||
-          (command is ChatCommand && command.type != CommandType.textOnly)) {
+          (command is ChatCommand && command.resolvedType != CommandType.textOnly)) {
         throw CommandRegistrationError('Cannot nest Slash commands!');
       }
     }